Added W

+++ crates/tools/sys/src/main.rs
@@ -3,22 +3,38 @@ use std::collections::*;
 use std::io::prelude::*;
 
 fn main() {
-    let rustfmt = std::env::args().nth(1).unwrap_or_default() != "-p";
+    let mut rustfmt = true;
+    let mut expect_namespace = false;
+    let mut namespace = String::new();
+    for arg in std::env::args() {
+        match arg.as_str() {
+            "-p" => rustfmt = false,
+            "-n" => expect_namespace = true,
+            _ => {
+                if expect_namespace {
+                    namespace = arg;
+                }
+            }
+        }
+    }
     let mut output = std::path::PathBuf::from("crates/libs/sys/src/Windows");
-    let _ = std::fs::remove_dir_all(&output);
+    if namespace.is_empty() {
+        let _ = std::fs::remove_dir_all(&output);
+    }
     output.pop();
-
     let files = vec![metadata::reader::File::new("crates/libs/metadata/default/Windows.winmd").unwrap(), metadata::reader::File::new("crates/libs/metadata/default/Windows.Win32.winmd").unwrap(), metadata::reader::File::new("crates/libs/metadata/default/Windows.Win32.Interop.winmd").unwrap()];
     let reader = &metadata::reader::Reader::new(&files);
+    if !namespace.is_empty() {
+        let tree = reader.tree(&namespace, &[]).expect("Namespace not found");
+        gen_tree(reader, &output, &tree, rustfmt);
+        return;
+    }
     let win32 = reader.tree("Windows.Win32", &lib::EXCLUDE_NAMESPACES).expect("`Windows.Win32` namespace not found");
     let root = metadata::reader::Tree { namespace: "Windows", nested: BTreeMap::from([("Win32", win32)]) };
-
     let trees = root.flatten();
     trees.par_iter().for_each(|tree| gen_tree(reader, &output, tree, rustfmt));
-
     output.pop();
     output.push("Cargo.toml");
-
     let mut file = std::fs::File::create(&output).unwrap();
 
     file.write_all(
@@ -99,7 +115,7 @@ fn gen_tree(reader: &metadata::reader::Reader, output: &std::path::Path, tree: &
     let mut path = std::path::PathBuf::from(output);
     path.push(tree.namespace.replace('.', "/"));
     std::fs::create_dir_all(&path).unwrap();
-
+    
     let mut gen = bindgen::Gen::new(reader);
     gen.namespace = tree.namespace;
     gen.sys = true;
